migrated to slash commands
