+++ events/ready.js
@@ -1,0 +1,24 @@
+import cron from 'node-cron'
+
+export default async (client, lc=client.lc) => {
+    console.log(`Logged in as ${client.user.tag}!`);
+
+    cron.schedule('0 6 * * *', async () => {
+        try {
+            const daily = await lc.daily();
+            const channel = client.channels.cache.get(process.env.CHANNEL_ID);
+            if (channel) {
+                const questionLink = `https://leetcode.com${daily.link}`;
+                const response = `@everyone **LeetCode Daily Challenge ${daily.date}:**\n**${daily.question.title}** : ${questionLink}`;
+                channel.send(response);
+            } else {
+                console.error('Channel not found');
+            }
+        } catch (error) {
+            console.error('Error fetching LeetCode daily challenge:', error);
+        }
+    }, {
+        scheduled: true,
+        timezone: "Asia/Kolkata"
+    });
+}
