+++ example.env
@@ -1,0 +1,3 @@
+TOKEN=
+CHANNEL_ID=
+DEVELOPER_ID=
